add actions add type move type definition from component
