+++ ts/features/wallet/onboarding/paypal/components/PspRadioItem.tsx
@@ -9,8 +9,8 @@ import { H4 } from "../../../../../components/core/typography/H4";
 import TouchableDefaultOpacity from "../../../../../components/TouchableDefaultOpacity";
 import IconFont from "../../../../../components/ui/IconFont";
 import { IOColors } from "../../../../../components/core/variables/IOColors";
-import { PayPalPsp } from "../screen/PayPalPspSelectionScreen";
 import { TestID } from "../../../../../types/WithTestID";
+import { PayPalPsp } from "../types";
 import { PspInfoBottomSheetContent } from "./PspInfoBottomSheet";
 
 const PSP_LOGO_MAX_WIDTH = Dimensions.get("window").width;
