+++ ts/features/wallet/onboarding/paypal/store/actions/index.ts
@@ -1,0 +1,12 @@
+import { createAsyncAction } from "typesafe-actions";
+import { NetworkError } from "../../../../../../utils/errors";
+import { PayPalPsp } from "../../types";
+
+/**
+ * Request the available psp for Paypal
+ */
+export const searchPaypalPsp = createAsyncAction(
+  "WALLET_ONBOARDING_PAYPAL_PSP_SEARCH_REQUEST",
+  "WALLET_ONBOARDING_PAYPAL_PSP_SEARCH_SUCCESS",
+  "WALLET_ONBOARDING_PAYPAL_PSP_SEARCH_FAILURE"
+)<void, ReadonlyArray<PayPalPsp>, NetworkError>();
